+++ EFCoreSecurity/DevExpress.EntityFramework.SecurityDataStore.Tests.Performance/Performance/DeleteTests.cs
@@ -74,8 +74,8 @@ namespace DevExpress.EntityFramework.SecurityDataStore.Tests.Performance {
                 }
             }
 
-            long securedContextTime = PerformanceTestsHelper.GetSecuredContextTime(times);
-            long nativeContextTime = PerformanceTestsHelper.GetNativeContextTime(times);
+            double securedContextTime = PerformanceTestsHelper.GetSecuredContextTime(times);
+            double nativeContextTime = PerformanceTestsHelper.GetNativeContextTime(times);
 
             Assert.IsTrue(false, "our: " + securedContextTime.ToString() + " ms, native: " + nativeContextTime.ToString() + " ms");
         }
